Re-organized code to better support background subtraction
