+++ jwst/cube_build/spaxel.py
@@ -1,0 +1,37 @@
+# Cube Class
+# Spaxel Class
+
+import sys
+import numpy as np
+import math
+import logging
+from .. import datamodels
+from . import coord
+
+log = logging.getLogger(__name__)
+log.setLevel(logging.DEBUG)
+
+
+
+##################################################################################
+class Spaxel(object):
+
+
+    __slots__ = ['flux', 'error','flux_weight','iflux']
+
+    def __init__(self):
+        self.flux = 0.0
+        self.flux_weight = 0.0
+        self.iflux = 0
+        self.error = 0
+
+class SpaxelAB(object):
+
+    __slots__ = ['flux', 'error', 'flux_weight','iflux']
+
+    def __init__(self):
+
+        self.flux = 0
+        self.error = 0
+        self.flux_weight = 0.0
+        self.iflux = 0.0
